+++ src/entities/persons/NaturalCustomer.java
@@ -7,7 +7,7 @@ public class NaturalCustomer extends Customer {
     private String rg;
     private String birthday;
 
-    public NaturalCustomer(Long id, String name, String zipCode, String telephone, String email, Double income, String username, String password, CustomerSituation customerSituation, String cpf, String rg, String birthday){
+    public NaturalCustomer(int id, String name, String zipCode, String telephone, String email, Double income, String username, String password, CustomerSituation customerSituation, String cpf, String rg, String birthday){
         super(id, name, zipCode, telephone, email, income, username, password, customerSituation);
         this.cpf = cpf;
         this.rg = rg;
